requirements.txt
requests==2.24.0

# testing
mock==3.0.5
httpretty==1.1.4
coveralls==1.11.1
coverage==5.2.1

# libraries
tiktoken==0.5.2
google-api-python-client==2.110.0
python-dotenv==1.0.0
youtube_transcript_api==0.6.1
openai==1.3.7
gradio==4.8.0